+++ client/src/Providers.tsx
@@ -3,15 +3,18 @@ import LoadingOverlayProvider from './providers/LoadingOverlayProvider';
 import PopupMessageProvider from './providers/PopupMessageProvider';
 import ConfirmModalProvider from './providers/ConfirmModalProvider';
 import InfoModalProvider from './providers/InfoModalProvider';
+import AuthProvider from './providers/AuthProvider';
 
 export default function Providers({ children }: { children: ReactNode }): JSX.Element {
   return (
-    <LoadingOverlayProvider>
-      <PopupMessageProvider>
-        <ConfirmModalProvider>
-          <InfoModalProvider>{children}</InfoModalProvider>
-        </ConfirmModalProvider>
-      </PopupMessageProvider>
-    </LoadingOverlayProvider>
+    <AuthProvider>
+      <LoadingOverlayProvider>
+        <PopupMessageProvider>
+          <ConfirmModalProvider>
+            <InfoModalProvider>{children}</InfoModalProvider>
+          </ConfirmModalProvider>
+        </PopupMessageProvider>
+      </LoadingOverlayProvider>
+    </AuthProvider>
   );
 }