+++ src/auth/authSessions.ts
@@ -2,12 +2,17 @@ import { Response } from 'express';
 import { dbPool } from '../db/db';
 import { PoolConnection, ResultSetHeader, RowDataPacket } from 'mysql2/promise';
 import { setResponseCookie } from '../util/cookieUtils';
-import { AUTH_SESSIONS_LIMIT, hourMilliseconds } from '../util/constants';
+import { AUTH_SESSIONS_LIMIT, dayMilliseconds, hourMilliseconds } from '../util/constants';
 import { generatePlaceHolders } from '../util/sqlUtils/generatePlaceHolders';
 import { isSqlError } from '../util/sqlUtils/isSqlError';
 import { generateCryptoUuid } from '../util/tokenGenerator';
 
-export async function createAuthSession(res: Response, accountId: number, keepSignedIn: boolean, attemptCount: number = 1): Promise<boolean> {
+export async function createAuthSession(
+  res: Response,
+  accountId: number,
+  keepSignedIn: boolean,
+  attemptCount: number = 1
+): Promise<boolean> {
   if (attemptCount > 3) {
     return false;
   }
@@ -15,8 +20,8 @@ export async function createAuthSession(res: Response, accountId: number, keepSi
   const newAuthSessionId: string = generateCryptoUuid();
   const currentTimestamp: number = Date.now();
 
-  const maxAge: number = keepSignedIn ? hourMilliseconds * 24 * 7 : hourMilliseconds * 6;
-  const expiryTimestamp: number = currentTimestamp + maxAge;
+  const maxAge: number | undefined = keepSignedIn ? dayMilliseconds * 7 : undefined;
+  const expiryTimestamp: number = currentTimestamp + (maxAge ? maxAge : hourMilliseconds * 6);
 
   let connection: PoolConnection | null = null;
 
@@ -48,9 +53,11 @@ export async function createAuthSession(res: Response, accountId: number, keepSi
           session_id,
           account_id,
           created_on_timestamp,
-          expiry_timestamp
-        ) VALUES (${generatePlaceHolders(4)});`,
-        [newAuthSessionId, accountId, currentTimestamp, expiryTimestamp]
+          expiry_timestamp,
+          keep_signed_in,
+          extensions_count
+        ) VALUES (${generatePlaceHolders(6)});`,
+        [newAuthSessionId, accountId, currentTimestamp, expiryTimestamp, keepSignedIn, 0]
       );
 
       await connection.commit();